add new leaderboard selectors
