+++ web-common/src/features/dashboards/dimension-table/DimensionHeader.svelte
@@ -38,8 +38,8 @@
       sorting: { sortedByDimensionValue },
       dimensions: { getDimensionDisplayName },
       dimensionFilters: { isFilterExcludeMode },
-      measures: {
-        visibleMeasures,
+      measures: { visibleMeasures },
+      leaderboard: {
         leaderboardShowAllMeasures,
         leaderboardMeasureNames,
         leaderboardSortByMeasureName,
