+++ web-common/src/features/dashboards/leaderboard/LeaderboardDisplay.svelte
@@ -18,6 +18,8 @@
   export let timeRange: V1TimeRange;
   export let comparisonTimeRange: V1TimeRange | undefined;
   export let timeControlsReady: boolean;
+
+  // FIXME: access selectors from store, instead of props
   export let leaderboardSortByMeasureName: string;
   export let leaderboardMeasureNames: string[];
 
@@ -29,11 +31,8 @@
       dimensions: { visibleDimensions },
       comparison: { isBeingCompared: isBeingComparedReadable },
       sorting: { sortedAscending, sortType, sortByMeasure },
-      measures: {
-        measureLabel,
-        isMeasureValidPercentOfTotal,
-        leaderboardShowAllMeasures,
-      },
+      measures: { measureLabel, isMeasureValidPercentOfTotal },
+      leaderboard: { leaderboardShowAllMeasures },
     },
     actions: {
       dimensions: { setPrimaryDimension },
